+++ lib/pesapal.rb
@@ -2,6 +2,7 @@ require "rails"
 require "htmlentities"
 require "pesapal/merchant"
 require "pesapal/merchant/post"
+require "pesapal/merchant/status"
 require "pesapal/oauth"
 require "pesapal/version"