+++ lib/pesapal/merchant.rb
@@ -72,7 +72,7 @@ module Pesapal
                 # required
                 @post_xml = Pesapal::Post::generate_post_xml @order_details
 
-                # initialize setting of @params (oauth_signature left empty) ... this gene
+                # initialize setting of @params (oauth_signature left empty)
                 @params = Pesapal::Post::set_parameters(@config[:callback_url], @config[:consumer_key], @post_xml)
 
                 # generate oauth signature and add signature to the request parameters
@@ -82,6 +82,26 @@ module Pesapal
                 query_string = Pesapal::Oauth::generate_encoded_params_query_string @params
 
                 "#{@api_endpoints[:postpesapaldirectorderv4]}?#{query_string}"
+            end
+
+            # query the status of the transaction
+            def query_payment_status(merchant_reference, transaction_tracking_id = nil)
+                
+                # initialize setting of @params (oauth_signature left empty)
+                @params = Pesapal::Status::set_parameters(@config[:consumer_key], merchant_reference, transaction_tracking_id)
+
+                # generate oauth signature and add signature to the request parameters
+                @params[:oauth_signature] = Pesapal::Oauth::generate_oauth_signature("GET", @api_endpoints[:querypaymentstatus], @params, @config[:consumer_secret], @token_secret)
+
+                # change params (with signature) to a query string
+                query_string = Pesapal::Oauth::generate_encoded_params_query_string @params
+
+                # get status response
+                response = Net::HTTP.get(URI("#{@api_endpoints[:querypaymentstatus]}?#{query_string}"))
+                response = CGI::parse(response)
+
+                # return the string result of what we want
+                response["pesapal_response_data"][0]
             end
 
         private
@@ -98,7 +118,6 @@ module Pesapal
                 @api_endpoints = {}
                 @api_endpoints[:postpesapaldirectorderv4] = "#{@api_domain}/API/PostPesapalDirectOrderV4"
                 @api_endpoints[:querypaymentstatus] = "#{@api_domain}/API/QueryPaymentStatus"
-                @api_endpoints[:querypaymentstatusbymerchantref] = "#{@api_domain}/API/QueryPaymentStatus"
                 @api_endpoints[:querypaymentdetails] = "#{@api_domain}/API/QueryPaymentDetails"
             end